snh-v2

+++ processbookappointment.php
@@ -1,6 +1,6 @@
 <?php session_start();
 //Collecting the data
-
+require_once('functions/user.php');
 
 
 $errorCount = 0;
@@ -14,13 +14,16 @@ $nature = $_POST['nature'] != "" ? $_POST['nature'] :  $errorCount++;
 $icomplaint = $_POST['icomplaint'] != "" ? $_POST['icomplaint'] :  $errorCount++;
 $department = $_POST['department'] != "" ? $_POST['department'] :  $errorCount++;
 
-
 $_SESSION['appdate'] = $appdate;
 $_SESSION['time'] = $time;
 $_SESSION['nature'] = $nature;
 $_SESSION['icomplaint'] = $icomplaint;
 $_SESSION['department'] = $department;
 
+if(isset($_SESSION['email'])){
+    $email= $_SESSION['email'];
+  }
+                
 
 if($errorCount > 0){
     $session_error = "You have " . $errorCount . " error";
@@ -45,33 +48,20 @@ if($errorCount > 0){
         'nature'=>$nature,
         'icomplaint'=>$icomplaint,
         'department'=>$department,
-
+        'email'=>$email,
 
     ];
+ 
+              $userExists = find_user($email);
 
-    //attempmt to create session and display on other pages remove if code misbehaves
-     // ($counter = 0; $counter < $countAllUsers ; $counter++) {
-       // $currentUser = $allUsers[$counter];
-       // if($currentUser == $email . ".json"){
-            //$AlluserString = file_get_contents("db/appointments/".$countAllUsers);
-            
-           // $userObject =json_decode($allUsers);
-            
-          //  $_SESSION["allUsers"] = $userObject -> id . " " . $userObject -> appdate . " " . $userObject -> time . " " . $userObject -> nature . " " . $userObject -> icomplaint . " " . $userObject -> department;
-            //ensure session is logged in vid 21
-            //$contents=array_map('file_get_contents',array_filter(glob('db/appointments/*'),function($file) {
-              //  return (is_file($file) && filesize($file)<30000);
-              //  }));    
-           
+              if($userExists){
+                file_put_contents("db/appointments/". $email . ".json", json_encode($userObject));
+                $_SESSION["message"] = "Appointment has been booked successfully. We will be expecting you ";
+                header("Location: appsuccess.php");
+                die();
+              }
 
-    
-
-    //save in the database;
-
-    file_put_contents("db/appointments/". $email . ".json", json_encode($userObject));
-
-    $_SESSION["message"] = "Appointment has been booked successfully. We will be expecting you ";
-
+    $_SESSION["error"] = "You are not suppossed to be on this page";
     header("Location: bookappointment.php");
     die();
